minor file re-adjustments. Preparations for deploying code the edge rpi3
