+++ pkg/ros/node.go
@@ -1,0 +1,6 @@
+package ros
+
+type Node struct {
+	Name      string
+	IPAddress string
+}
