+++ push_and_build.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+IP_ADDRESS=$1
+HOSTNAME=$2
+
